pyproject.toml
[tool.poetry]
name = "gh-utils"
version = "0.1.4"
description = "GitHub CLI Utilities"
authors = ["Xinyuan Chen <45612704+tddschn@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/tddschn/gh-utils"
repository = "https://github.com/tddschn/gh-utils"
classifiers = [
    "Topic :: Utilities"
]
keywords = ["gh", "github", "utils"]

[tool.poetry.scripts]
ghcrar = "gh_utils.gh_create_repo_and_add_to_remote:main"

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/tddschn/gh-utils/issues"

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0"

[tool.poetry.group.dev.dependencies]
ipython = "^8.4.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.3.0"
better-exceptions = "^0.3.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
